% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/contents.R
\name{contents}
\alias{contents}
\title{Basic summary of the data}
\usage{
contents(data)
}
\arguments{
\item{data}{a data.frame or data.table}
}
\value{
a data.frame that contains variable, class, distinct values,
         missing values, percentage of missing value and sample value
}
\description{
Takes in a data and returns summary of the data
}
\details{
This function helps when one wants to get a quick snapshot of the data such as class,
 distinct values, missing values and sample value of the variables.

 It works for both 'data.frame' and 'data.table' but the output will be a 'data.frame' only.
}
\examples{
# A data frame
df <- data.frame(x = c(1, 2, 3, 4, NA),
                 y = c('a', 'b', 'c', 'd', 'e'),
                 z = c(1, 1, 0, 0, 1))

# Summary of the data
dfContents <- contents(data = df)
}
\author{
Akash Jain
}
\seealso{
\code{\link{dqcontinuous}}, \code{\link{dqcategorical}}, \code{\link{dqdate}}
}

